recommendation-ml/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
mysql-connector-python==9.1.0
pandas>=2.0,<3
numpy>=1.26.4,<2.1
scikit-learn>=1.4,<1.6
pydantic==2.10.3
pydantic-settings==2.6.1
pytest==8.3.4
